refactor: 패키지 이름, 구조 정리
