+++ src/main/java/domain/menu/Menu.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.menu;
+
+import domain.Category;
 
 public class Menu {
     private final int number;
