+++ src/main/java/domain/menu/MenuRepository.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.menu;
+
+import domain.Category;
 
 import java.util.ArrayList;
 import java.util.Collections;
