+++ src/main/java/domain/order/Order.java
@@ -1,0 +1,4 @@
+package domain.order;
+
+public class Order {
+}
